# swagger-ui.json
{"openapi":"3.1.0","info":{"title":"Sammelalbum API","description":"API for managing sticker collections and searches","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Matches","description":"Endpoints for finding potential card exchange matches"},{"name":"Card Offers","description":"Endpoints for managing card offers"},{"name":"Card Search","description":"Endpoints for managing card searches"},{"name":"Exchanges","description":"Endpoints for managing card exchange requests"},{"name":"Authentication","description":"Endpoints for user registration, login, and password reset"},{"name":"User","description":"Endpoints for user management"}],"paths":{"/api/user/profile":{"put":{"tags":["User"],"summary":"Update user profile","description":"Updates the profile information of the currently authenticated user","operationId":"updateProfile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProfileRequest"}}},"required":true},"responses":{"200":{"description":"Profile updated successfully"},"403":{"description":"User not authorized"}},"security":[{"bearerAuth":[]}]}},"/api/user/password":{"put":{"tags":["User"],"summary":"Change password","description":"Changes the password of the currently authenticated user","operationId":"changePassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordRequest"}}},"required":true},"responses":{"200":{"description":"Password changed successfully"},"400":{"description":"Invalid current password"},"403":{"description":"User not authorized"}},"security":[{"bearerAuth":[]}]}},"/api/exchanges/{id}/decline":{"put":{"tags":["Exchanges"],"summary":"Decline exchange request","description":"Declines an exchange request and unreserves any reserved cards","operationId":"declineExchangeRequest","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Exchange request declined successfully"},"400":{"description":"Invalid request or exchange not in correct state"},"403":{"description":"User not authorized to decline this exchange"}},"security":[{"bearerAuth":[]}]}},"/api/exchanges/{id}/close":{"put":{"tags":["Exchanges"],"summary":"Close exchange request","description":"Marks an exchange as completed and deletes the exchanged cards","operationId":"closeExchangeRequest","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Exchange request closed successfully"},"400":{"description":"Invalid request or exchange not in accepted state"},"403":{"description":"User not authorized to close this exchange"}},"security":[{"bearerAuth":[]}]}},"/api/exchanges/{id}/accept":{"put":{"tags":["Exchanges"],"summary":"Accept exchange request","description":"Accepts an exchange request and reserves the involved cards","operationId":"acceptExchangeRequest","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Exchange request accepted successfully"},"400":{"description":"Invalid request or exchange already accepted/declined"},"403":{"description":"User not authorized to accept this exchange"}},"security":[{"bearerAuth":[]}]}},"/api/card-offers/bulk":{"put":{"tags":["Card Offers"],"summary":"Update multiple offers","description":"Updates the offer type for all offers matching the given sticker IDs for the authenticated user","operationId":"updateBulkOffers","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BulkUpdateOfferRequest"}}},"required":true},"responses":{"200":{"description":"Card offers updated successfully","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CardOfferResponse"}}}}},"400":{"description":"Invalid input","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CardOfferResponse"}}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["Card Offers"],"summary":"Add multiple offers","description":"Adds multiple card offers for the authenticated user","operationId":"addBulkOffers","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BulkCardOfferRequest"}}},"required":true},"responses":{"200":{"description":"Card offers added successfully","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CardOfferResponse"}}}}},"400":{"description":"Invalid input","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CardOfferResponse"}}}}}},"security":[{"bearerAuth":[]}]}},"/api/exchanges":{"post":{"tags":["Exchanges"],"summary":"Create exchange request","description":"Creates a new exchange request between the authenticated user and another user","operationId":"createExchangeRequest","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateExchangeRequestDto"}}},"required":true},"responses":{"200":{"description":"Exchange request created successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExchangeRequest"}}}},"400":{"description":"Invalid request (e.g., requesting from self, invalid stickers, or exchange type mismatch)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExchangeRequest"}}}}},"security":[{"bearerAuth":[]}]}},"/api/card-searches":{"get":{"tags":["Card Search"],"summary":"Get all searches","description":"Retrieves all card searches for the authenticated user","operationId":"getSearches","responses":{"200":{"description":"List of card searches retrieved successfully","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CardSearchResponse"}}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["Card Search"],"summary":"Add a search","description":"Adds a single card search for the authenticated user","operationId":"addSearch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CardSearchRequest"}}},"required":true},"responses":{"200":{"description":"Card search added successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardSearchResponse"}}}},"400":{"description":"Invalid input","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardSearchResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/card-searches/bulk":{"post":{"tags":["Card Search"],"summary":"Add multiple searches","description":"Adds multiple card searches for the authenticated user","operationId":"addBulkSearches","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BulkCardSearchRequest"}}},"required":true},"responses":{"200":{"description":"Card searches added successfully","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CardSearchResponse"}}}}},"400":{"description":"Invalid input","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CardSearchResponse"}}}}}},"security":[{"bearerAuth":[]}]}},"/api/card-searches/bulk-delete":{"post":{"tags":["Card Search"],"summary":"Delete multiple searches","description":"Deletes multiple card searches for the authenticated user, handling duplicates intelligently","operationId":"removeBulkSearches","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BulkCardSearchRequest"}}},"required":true},"responses":{"204":{"description":"Card searches deleted successfully"}},"security":[{"bearerAuth":[]}]}},"/api/card-offers":{"get":{"tags":["Card Offers"],"summary":"Get all offers","description":"Retrieves all card offers for the authenticated user","operationId":"getOffers","responses":{"200":{"description":"List of card offers retrieved successfully","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CardOfferResponse"}}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["Card Offers"],"summary":"Add an offer","description":"Adds a single card offer for the authenticated user","operationId":"addOffer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CardOfferRequest"}}},"required":true},"responses":{"200":{"description":"Card offer added successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardOfferResponse"}}}},"400":{"description":"Invalid input","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardOfferResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/card-offers/bulk-delete":{"post":{"tags":["Card Offers"],"summary":"Delete multiple offers","description":"Deletes multiple card offers for the authenticated user, handling duplicates intelligently","operationId":"removeBulkOffers","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BulkCardOfferRequest"}}},"required":true},"responses":{"204":{"description":"Card offers deleted successfully"}},"security":[{"bearerAuth":[]}]}},"/api/auth/reset-password":{"post":{"tags":["Authentication"],"summary":"Reset password","description":"Resets user password using a valid password reset token","operationId":"resetPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordRequest"}}},"required":true},"responses":{"200":{"description":"Password reset successfully"},"400":{"description":"Invalid or expired token"}}}},"/api/auth/register":{"post":{"tags":["Authentication"],"summary":"Register a new user","description":"Creates a new user account and returns a JWT token","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"User registered successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"400":{"description":"Invalid input or validation failed","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/auth/login":{"post":{"tags":["Authentication"],"summary":"Login user","description":"Authenticates a user and returns a JWT token","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"User logged in successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"403":{"description":"Invalid credentials","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/auth/forgot-password":{"post":{"tags":["Authentication"],"summary":"Request password reset","description":"Initiates password reset process by sending an email with a reset link. Always returns 204 for security reasons, regardless of whether the user exists.","operationId":"forgotPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordRequest"}}},"required":true},"responses":{"204":{"description":"Password reset email sent if user exists"},"400":{"description":"Invalid input"}}}},"/api/user/me":{"get":{"tags":["User"],"summary":"Get current user","description":"Retrieves the profile of the currently authenticated user","operationId":"getMe","responses":{"200":{"description":"User profile retrieved successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"403":{"description":"User not authorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["User"],"summary":"Delete current user","description":"Permanently deletes the authenticated user's account and all associated data. This action cannot be undone.","operationId":"deleteMe","responses":{"204":{"description":"User deleted successfully"},"403":{"description":"User not authorized"}},"security":[{"bearerAuth":[]}]}},"/api/statistics":{"get":{"tags":["statistics-controller"],"operationId":"getStatistics","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StatisticsResponse"}}}}}}},"/api/matches/payed":{"get":{"tags":["Matches"],"summary":"Get paid matches","description":"Retrieves paginated list of users offering cards for payment that match the authenticated user's searches","operationId":"getPayedMatches","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageMatchResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/matches/freebie":{"get":{"tags":["Matches"],"summary":"Get freebie matches","description":"Retrieves paginated list of users offering cards for free that match the authenticated user's searches","operationId":"getFreebieMatches","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageMatchResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/matches/exchange":{"get":{"tags":["Matches"],"summary":"Get exchange matches","description":"Retrieves paginated list of potential card-for-card exchanges with other users","operationId":"getExchangeMatches","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageMatchResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/exchanges/sent":{"get":{"tags":["Exchanges"],"summary":"Get sent requests","description":"Retrieves all exchange requests sent by the authenticated user","operationId":"getSentRequests","responses":{"200":{"description":"List of sent exchange requests","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ExchangeRequestDto"}}}}}},"security":[{"bearerAuth":[]}]}},"/api/exchanges/received":{"get":{"tags":["Exchanges"],"summary":"Get received requests","description":"Retrieves all exchange requests received by the authenticated user","operationId":"getReceivedOffers","responses":{"200":{"description":"List of received exchange requests","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ExchangeRequestDto"}}}}}},"security":[{"bearerAuth":[]}]}},"/api/card-searches/{id}":{"delete":{"tags":["Card Search"],"summary":"Delete a search","description":"Deletes a specific card search by ID","operationId":"deleteSearch","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"Card search deleted successfully"},"404":{"description":"Card search not found or not owned by user"}},"security":[{"bearerAuth":[]}]}},"/api/card-offers/{id}":{"delete":{"tags":["Card Offers"],"summary":"Delete an offer","description":"Deletes a specific card offer by ID","operationId":"deleteOffer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"Card offer deleted successfully"},"404":{"description":"Card offer not found or not owned by user"}},"security":[{"bearerAuth":[]}]}}},"components":{"schemas":{"UpdateProfileRequest":{"type":"object","properties":{"firstname":{"type":"string"},"lastname":{"type":"string"},"mail":{"type":"string"},"contact":{"type":"string"}}},"ChangePasswordRequest":{"type":"object","properties":{"currentPassword":{"type":"string"},"newPassword":{"type":"string"}}},"BulkUpdateOfferRequest":{"type":"object","properties":{"stickerIds":{"type":"array","items":{"type":"integer","format":"int64"}},"offerPayed":{"type":"boolean"},"offerFreebie":{"type":"boolean"},"offerExchange":{"type":"boolean"}}},"CardOfferResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"stickerId":{"type":"integer","format":"int64"},"stickerName":{"type":"string"},"offerPayed":{"type":"boolean"},"offerFreebie":{"type":"boolean"},"offerExchange":{"type":"boolean"},"isReserved":{"type":"boolean"}}},"CreateExchangeRequestDto":{"type":"object","properties":{"offererId":{"type":"integer","format":"int64"},"requestedStickerId":{"type":"integer","format":"int64"},"offeredStickerId":{"type":"integer","format":"int64"},"exchangeType":{"type":"string","enum":["EXCHANGE","PAYED","FREEBIE"]}}},"ExchangeRequest":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"requesterId":{"type":"integer","format":"int64"},"offererId":{"type":"integer","format":"int64"},"requestedStickerId":{"type":"integer","format":"int64"},"offeredStickerId":{"type":"integer","format":"int64"},"exchangeType":{"type":"string","enum":["EXCHANGE","PAYED","FREEBIE"]},"status":{"type":"string","enum":["INITIAL","MAIL_SEND","EXCHANGE_INTERREST","EXCHANGE_COMPLETED","EXCHANGE_CANCELED"]},"offererCardOfferId":{"type":"integer","format":"int64"},"requesterCardSearchId":{"type":"integer","format":"int64"},"requesterCardOfferId":{"type":"integer","format":"int64"},"offererCardSearchId":{"type":"integer","format":"int64"},"requesterClosed":{"type":"boolean"},"offererClosed":{"type":"boolean"},"cancellationReason":{"type":"string","enum":["REQUESTER_CANCELED","OFFERER_CANCELED","OFFERED_CARD_REMOVED_BY_USER","SEARCH_CARD_REMOVED_BY_USER","USER_DELETION"]},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"CardSearchRequest":{"type":"object","properties":{"stickerId":{"type":"integer","format":"int64"}}},"CardSearchResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"stickerId":{"type":"integer","format":"int64"},"stickerName":{"type":"string"},"isReserved":{"type":"boolean"}}},"BulkCardSearchRequest":{"type":"object","properties":{"stickerIds":{"type":"array","items":{"type":"integer","format":"int64"}}}},"CardOfferRequest":{"type":"object","properties":{"stickerId":{"type":"integer","format":"int64"},"offerPayed":{"type":"boolean"},"offerFreebie":{"type":"boolean"},"offerExchange":{"type":"boolean"}}},"BulkCardOfferRequest":{"type":"object","properties":{"stickerIds":{"type":"array","items":{"type":"integer","format":"int64"}},"offerPayed":{"type":"boolean"},"offerFreebie":{"type":"boolean"},"offerExchange":{"type":"boolean"}}},"ResetPasswordRequest":{"type":"object","description":"Request to reset password using a valid token","properties":{"token":{"type":"string","description":"Password reset JWT token","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."},"newPassword":{"type":"string","description":"New password (minimum 8 characters)","example":"NewSecurePass123!","maxLength":2147483647,"minLength":8}},"required":["newPassword","token"]},"RegisterRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"mail":{"type":"string"},"firstname":{"type":"string"},"lastname":{"type":"string"},"contact":{"type":"string"},"validationCode":{"type":"string"}},"required":["firstname","lastname","mail","password","username","validationCode"]},"AuthResponse":{"type":"object","properties":{"token":{"type":"string"},"message":{"type":"string"},"userId":{"type":"integer","format":"int64"}}},"LoginRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["password","username"]},"ForgotPasswordRequest":{"type":"object","description":"Request to initiate password reset process","properties":{"identifier":{"type":"string","description":"Username or email address","example":"john.doe@example.com"}},"required":["identifier"]},"UserDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstname":{"type":"string"},"lastname":{"type":"string"},"mail":{"type":"string"},"contact":{"type":"string"}}},"CardOfferStatistics":{"type":"object","properties":{"total":{"type":"integer","format":"int64"},"free":{"type":"integer","format":"int64"},"exchange":{"type":"integer","format":"int64"},"paid":{"type":"integer","format":"int64"}}},"CardSearchStatistics":{"type":"object","properties":{"total":{"type":"integer","format":"int64"}}},"StatisticsResponse":{"type":"object","properties":{"cardOffers":{"$ref":"#/components/schemas/CardOfferStatistics"},"cardSearches":{"$ref":"#/components/schemas/CardSearchStatistics"}}},"Pageable":{"type":"object","properties":{"page":{"type":"integer","format":"int32","minimum":0},"size":{"type":"integer","format":"int32","minimum":1},"sort":{"type":"array","items":{"type":"string"}}}},"MatchResponse":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"exchangeableCount":{"type":"integer","format":"int64"},"itemsRequested":{"type":"array","items":{"$ref":"#/components/schemas/MatchStickerDto"}},"itemsOffered":{"type":"array","items":{"$ref":"#/components/schemas/MatchStickerDto"}}}},"MatchStickerDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"PageMatchResponse":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"numberOfElements":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"sort":{"$ref":"#/components/schemas/SortObject"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/MatchResponse"}},"number":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"sort":{"$ref":"#/components/schemas/SortObject"},"paged":{"type":"boolean"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"unpaged":{"type":"boolean"},"offset":{"type":"integer","format":"int64"}}},"SortObject":{"type":"object","properties":{"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"},"empty":{"type":"boolean"}}},"ExchangeRequestDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"requesterId":{"type":"integer","format":"int64"},"offererId":{"type":"integer","format":"int64"},"requestedStickerId":{"type":"integer","format":"int64"},"offeredStickerId":{"type":"integer","format":"int64"},"exchangeType":{"type":"string","enum":["EXCHANGE","PAYED","FREEBIE"]},"status":{"type":"string","enum":["INITIAL","MAIL_SEND","EXCHANGE_INTERREST","EXCHANGE_COMPLETED","EXCHANGE_CANCELED"]},"cancellationReason":{"type":"string","enum":["REQUESTER_CANCELED","OFFERER_CANCELED","OFFERED_CARD_REMOVED_BY_USER","SEARCH_CARD_REMOVED_BY_USER","USER_DELETION"]},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"requesterClosed":{"type":"boolean"},"offererClosed":{"type":"boolean"},"requestedStickerName":{"type":"string"},"offeredStickerName":{"type":"string"},"partnerFirstname":{"type":"string"},"partnerLastname":{"type":"string"},"partnerContact":{"type":"string"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}